Layout arrange fixes
